requirements.txt
et_xmlfile==2.0.0
lxml==5.3.1
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
six==1.17.0
typing_extensions==4.13.0
tzdata==2025.2
